file check

+++ setup.sh
@@ -23,11 +23,15 @@ setting_ctf() {
         $SUDO chmod 6555 $BIN
         $SUDO chown root:root $SRC
         $SUDO chmod 644 $SRC
-        $SUDO cp ~/.gdbinit /home/$USER/
-        $SUDO sed -i 's@source ~@source '"$HOME"'@g' /home/$USER/.gdbinit
-        $SUDO chown $USER:$USER /home/$USER/.gdbinit
-        $SUDO cp ~/.tmux.conf /home/$USER/
-        $SUDO chown $USER:$USER /home/$USER/.tmux.conf
+        if [[ -f ~/.gdbinit ]]; then
+            $SUDO cp ~/.gdbinit /home/$USER/
+            $SUDO sed -i 's@source ~@source '"$HOME"'@g' /home/$USER/.gdbinit
+            $SUDO chown $USER:$USER /home/$USER/.gdbinit
+        fi
+        if [[ -f ~/.tmux.conf ]]; then
+            $SUDO cp ~/.tmux.conf /home/$USER/
+            $SUDO chown $USER:$USER /home/$USER/.tmux.conf
+        fi
         $SUDO cp aslr /home/$USER/.aslr
         $SUDO chown root:root /home/$USER/.aslr
         $SUDO chmod 4775 /home/$USER/.aslr
